+++ src/pdftl/cli/help_data.py
@@ -18,7 +18,7 @@ License: MPL-2.0 <https://mozilla.org/MPL/2.0/>
 This is free software: you are free to change and redistribute it.
 There is NO WARRANTY, to the extent permitted by law.
 
-Core dependencies (and installed versions):
+Core dependencies:
 {dependencies}
 """