+++ tests/utils/test_string.py
@@ -1,5 +1,3 @@
-import decimal
-import sys
 from unittest.mock import MagicMock, patch
 
 import pytest